requirements.txt
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
# TODO: Prevent mysql-connector-python from installing its own libcrypto.so and
# libssl.so, see https://bugs.mysql.com/bug.php?id=89889.
#--install-option='--no-deps' --install-option='--only-binary=:all:'
mysql-connector-python==8.0.26
protobuf==3.17.3
PyQt5==5.14.2; platform_system != 'Windows'
PyQt5==5.15.6; platform_system == 'Windows'
PyQt5-sip==12.9.0
requests==2.26.0
six==1.16.0
urllib3==1.26.6
